+++ TriblerGUI/widgets/network_explorer/help_page.css
