+++ TriblerGUI/widgets/network_explorer/index.html
@@ -2,7 +2,7 @@
 <html lang="en">
 <head>
     <meta charset="UTF-8">
-    <title>Trust Page</title>
+    <title>Network Explorer</title>
     <script src="js/d3/d3.v4.min.js"></script>
     <link rel="stylesheet" type="text/css" href="style.css">
     <link rel="stylesheet" type="text/css" href="help_page.css">
@@ -19,7 +19,7 @@
 
 <div id="help-page">
     <div id="help-page-content">
-        <h2 id="help-page-title">Visualization of the Tribler network</h2>
+        <h2 id="help-page-title">Visualization of the Tribler Network</h2>
         <p>In the Tribler network, users exchange content anonymously by uploading and downloading. This only works if
             everyone makes content available by uploading. Users that decide not to upload at least as much as they
             download, are blocked from downloading until they start uploading again.</p>
