+++ TriblerGUI/widgets/network_explorer/js/help_page.js
