+++ TriblerGUI/widgets/network_explorer/js/request_data.js
