+++ TriblerGUI/widgets/network_explorer/js/style_config.js
