+++ TriblerGUI/widgets/network_explorer/js/support/eventful_mixin.js
