+++ TriblerGUI/widgets/network_explorer/style.css
