+++ TriblerGUI/widgets/networkexplorerpage.py
@@ -1,5 +1,7 @@
 """
-Provides a class which initializes the Network Trust Display Qt elements.
+Provides a class which initializes a page with information about the Tribler network.
+This is either a Network Explorer if the QWebEngineView import is available on the running platform or
+a trust graph if it is not.
 """
 import os
 from PyQt5.QtCore import QUrl
@@ -7,24 +9,24 @@ from PyQt5.QtGui import QColor
 from PyQt5.QtWidgets import QWidget
 from TriblerGUI import utilities
 
-TRUST_NETWORK_HTML_PATH = os.path.join(utilities.get_base_path(), "widgets/trustpage/index.html")
+NETWORK_EXPLORER_HTML_PATH = os.path.join(utilities.get_base_path(), "widgets/network_explorer/index.html")
 
 
-class TrustPageNetwork(QWidget):
+class NetworkExplorerPage(QWidget):
     """
-    The logic of the Network Trust Display.
+    The logic of the Network Explorer.
     """
 
     def __init__(self):
         """
-        Create a new Network Trust Display.
+        Create a new Network Explorer.
         """
         QWidget.__init__(self)
         self.network_graph = None
 
-    def initialize_trust_page(self):
+    def initialize_web_page(self):
         """
-        Load the pyplot graph into the QWidget.
+        Load the web page the QWidget.
         """
         try:
             from PyQt5.QtWebEngineWidgets import QWebEngineView
@@ -32,11 +34,11 @@ class TrustPageNetwork(QWidget):
 
             view = QWebEngineView()
 
-            view.setUrl(QUrl.fromLocalFile(TRUST_NETWORK_HTML_PATH))
+            view.setUrl(QUrl.fromLocalFile(NETWORK_EXPLORER_HTML_PATH))
             view.page().setBackgroundColor(QColor.fromRgb(0, 0, 0, 0))
             view.show()
 
             vertical_layout.addWidget(view)
         except ImportError:
-            # In the case QWebEngineView is not available, render the graph trust page.
+            # In the case QWebEngineView is not available, render the trust graph page.
             pass
